RecyclerItem defined and RecyclerItemAdapter to handle RecyclerItems
